refactor : ViewConstants, ExceptionConstants 패키지화
