+++ src/main/java/racingcar/constant/ExceptionConstants.java
@@ -1,0 +1,9 @@
+package racingcar.constant;
+
+public class ExceptionConstants {
+
+  public static final String UNALLOWED_SPECIAL_SYMBOL = "문자열에 특수 문자가 포함되어 있습니다.";
+  public static final String INVALID_SCOPE_NAME = "자동차 이름은 1~5자 이여야 합니다.";
+  public static final String OVER_INTMAX_ROUNDCOUNT = "라운드 횟수는 INT_MAXVALUE보다 클 수 없습니다.";
+  public static final String NEGATIVE_NUMBER_ROUNDCOUNT = "라운드 횟수는 0보다 작거나 같을 수 없습니다.";
+}
