+++ src/main/java/racingcar/constant/ViewConstants.java
@@ -1,9 +1,7 @@
-package racingcar.view;
+package racingcar.constant;
 
-public class ViewNotice {
-
+public class ViewConstants {
   public static final String GET_CARNAME_NOTICE = "경주할 자동차 이름을 입력하세요.(이름은 쉼표(,) 기준으로 구분)";
   public static final String GET_ROUND_NOTICE = "시도할 회수는 몇회인가요?";
   public static final String ROUND_RESULT_NOTICE = "실행 결과";
-
 }
