+++ src/main/java/racingcar/validator/RacingCarNameValidator.java
@@ -1,5 +1,8 @@
 package racingcar.validator;
 
+import static racingcar.constant.ExceptionConstants.INVALID_SCOPE_NAME;
+import static racingcar.constant.ExceptionConstants.UNALLOWED_SPECIAL_SYMBOL;
+
 import racingcar.exception.RacingCarGameException;
 import racingcar.model.RacingCarGameConfig;
 
@@ -13,13 +16,13 @@ public class RacingCarNameValidator implements Validator<String> {
 
   private void validateElement(final String data) {
     if (data.matches(".*[!@#$%^&*()_+\\-=\\[\\]{};':\"\\\\|,.<>/?].*")) {
-      throw new RacingCarGameException("문자열에 특수 문자가 포함되어 있습니다.");
+      throw new RacingCarGameException(UNALLOWED_SPECIAL_SYMBOL);
     }
   }
 
   private void validateLength(final String data) {
     if (data.length() > RacingCarGameConfig.CARNAME_MAXIMUM_LENGTH || data.isBlank()) {
-      throw new RacingCarGameException("자동차 이름은 1~5자 이여야 합니다.");
+      throw new RacingCarGameException(INVALID_SCOPE_NAME);
     }
   }
